fix(admin-dashboard): keep billing format helpers client-safe
Move formatBillingPeriod/Wei into a Node-free module so AdminUsagePanel no longer pulls OpenMeter/postgres into the client webpack graph.
