+++ src/components/DashboardUsagePanel.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import UsageLineChart from "@/components/UsageLineChart";
-import { formatBillingPeriod } from "@/lib/billing-usage-dashboard-data";
+import { formatBillingPeriod } from "@/lib/billing-format";
 import { getDashboardUsageSummary } from "@/lib/dashboard-usage-summary";
 import { formatUsdMicrosString } from "@/lib/format-usd-micros";
