+++ src/lib/billing-format.ts
@@ -1,0 +1,24 @@
+/** Client-safe billing display helpers (no Node/DB imports). */
+
+export function formatBillingWei(wei: string): string {
+  if (!wei || !/^\d+$/.test(wei)) return "0";
+  const value = BigInt(wei);
+  if (value === 0n) return "0";
+  const divisor = 10n ** 18n;
+  const whole = value / divisor;
+  const remainder = value % divisor;
+  if (whole === 0n && remainder > 0n) return `${value.toString()} wei`;
+  const fracStr = remainder.toString().padStart(18, "0").slice(0, 6);
+  return `${whole}.${fracStr} ETH`;
+}
+
+export function formatBillingPeriod(iso: string): string {
+  try {
+    return new Date(iso).toLocaleString(undefined, {
+      dateStyle: "medium",
+      timeStyle: "short",
+    });
+  } catch {
+    return iso;
+  }
+}
